refactor: delete executeScreenshotsForAllStates utility
