+++ packages/sit-onyx/src/components/OnyxBadge/OnyxBadge.ct.tsx
@@ -1,8 +1,7 @@
 import { DENSITIES } from "../../composables/density";
 import { expect, test } from "../../playwright-axe";
-import { executeMatrixScreenshotTest } from "../../playwright/screenshots";
+import { executeMatrixScreenshotTest, mockPlaywrightIcon } from "../../playwright/screenshots";
 import { ONYX_COLORS } from "../../types/colors";
-import { mockPlaywrightIcon } from "../../utils/playwright";
 import OnyxBadge from "./OnyxBadge.vue";
 
 test.describe("Screenshot tests", () => {
