+++ packages/sit-onyx/src/components/OnyxCheckboxGroup/OnyxCheckboxGroup.ct.tsx
@@ -1,4 +1,3 @@
-import { executeScreenshotsForAllStates } from "../../utils/playwright";
 import { expect, test } from "../../playwright-axe";
 import { DIRECTIONS } from "../../types";
 import OnyxCheckboxGroup from "./OnyxCheckboxGroup.vue";
@@ -110,38 +109,34 @@ test("should disabled all checkboxes if group is disabled", async ({ mount }) =>
   }
 });
 
-const STATES = {
-  state: ["required", "optional"],
-} as const;
-test.describe("should truncate", () => {
-  const options: OnyxCheckboxGroupProps["options"] = [
-    { label: "Very long label that will be truncated", id: "id-1" },
-    { label: "Very long required label that will be truncated", id: "id-2", required: true },
-    {
-      label: "Very long label that will be wrapped with multiline",
-      id: "id-3",
-      truncation: "multiline",
-    },
-    {
-      label: "Very long required label that will be wrapped with multiline",
-      id: "id-4",
-      truncation: "multiline",
-      required: true,
-    },
-  ];
-  executeScreenshotsForAllStates(STATES, "truncated-checkbox-group", async ({ state }, mount) => {
+for (const state of ["required", "optional"] as const) {
+  test(`should truncate (${state})`, async ({ mount }) => {
     const component = await mount(
       <OnyxCheckboxGroup
-        options={options}
+        options={[
+          { label: "Very long label that will be truncated", id: "id-1" },
+          { label: "Very long required label that will be truncated", id: "id-2", required: true },
+          {
+            label: "Very long label that will be wrapped with multiline",
+            id: "id-3",
+            truncation: "multiline",
+          },
+          {
+            label: "Very long required label that will be wrapped with multiline",
+            id: "id-4",
+            truncation: "multiline",
+            required: true,
+          },
+        ]}
         headline="Truncated group headline"
-        style="max-width: 16rem;"
+        class={{ "onyx-use-optional": state === "optional" }}
+        style={{ maxWidth: "16rem" }}
       />,
-      { useOptional: state === "optional" },
     );
 
-    return component;
+    await expect(component).toHaveScreenshot(`truncation-${state}.png`);
   });
-});
+}
 
 DIRECTIONS.forEach((direction) => {
   test(`should render ${direction} skeletons`, async ({ mount }) => {
