+++ packages/sit-onyx/src/components/OnyxEmpty/OnyxEmpty.ct.tsx
@@ -1,5 +1,5 @@
 import { expect, test } from "../../playwright-axe";
-import { mockPlaywrightIcon } from "../../utils/playwright";
+import { mockPlaywrightIcon } from "../../playwright/screenshots";
 import OnyxIcon from "../OnyxIcon/OnyxIcon.vue";
 import OnyxEmpty from "./OnyxEmpty.vue";
