+++ packages/sit-onyx/src/components/OnyxIcon/OnyxIcon.ct.tsx
@@ -1,7 +1,6 @@
 import { test } from "../../playwright-axe";
-import { executeMatrixScreenshotTest } from "../../playwright/screenshots";
+import { executeMatrixScreenshotTest, mockPlaywrightIcon } from "../../playwright/screenshots";
 import { ONYX_COLORS } from "../../types/colors";
-import { mockPlaywrightIcon } from "../../utils/playwright";
 import OnyxIcon from "./OnyxIcon.vue";
 import { ICON_SIZES } from "./types";
