+++ packages/sit-onyx/src/playwright/screenshots.tsx
@@ -108,3 +108,12 @@ export const executeMatrixScreenshotTest = async <TColumn extends string, TRow e
     await expect(component).toHaveScreenshot(`${options.name}.png`);
   });
 };
+
+/**
+ * Mock icon to use in Playwright component tests (.tsx files) because Playwright has
+ * issues when importing from "@sit-onyx/icons" directly with version 1.42.1.
+ *
+ * Equivalent to:
+ * import mockPlaywrightIcon from "@sit-onyx/icons/emoji-happy-2.svg?raw";
+ */
+export const mockPlaywrightIcon = `<svg xmlns="http://www.w3.org/2000/svg" xml:space="preserve" viewBox="0 0 32 32"><path d="M16 2C8.28 2 2 8.28 2 16s6.28 14 14 14 14-6.28 14-14S23.72 2 16 2m0 26C9.383 28 4 22.617 4 16S9.383 4 16 4s12 5.383 12 12-5.383 12-12 12"/><path d="M11 13c.552 0 1 .449 1 1h2c0-1.654-1.346-3-3-3s-3 1.346-3 3h2c0-.551.448-1 1-1m10-2c-1.654 0-3 1.346-3 3h2a1.001 1.001 0 0 1 2 0h2c0-1.654-1.346-3-3-3m-5 11a5.01 5.01 0 0 1-4.325-2.501l-1.73 1.002C11.193 22.659 13.514 24 16 24s4.807-1.341 6.056-3.499l-1.73-1.002A5.02 5.02 0 0 1 16 22"/></svg>`;
